requirements.txt
streamlit==1.31.1
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
plotly==5.18.0
numpy==1.24.3 